[kernel/libc] Rename posix_types.h to posixtyp.h for DEV86 build on FAT
